+++ packages/stream_video_flutter/test/src/call_screen/call_content/call_content_test.dart
@@ -31,6 +31,19 @@ void main() {
         final filtered = selector(mockCallState);
         return Stream.value(filtered);
       });
+      when(
+        () => mockCall
+            .partialState<({bool isScreenShareEnabled, CallStatus status})>(
+              any(),
+            ),
+      ).thenAnswer((invocation) {
+        return Stream.value(
+          (
+            isScreenShareEnabled: false,
+            status: CallStatus.connected(),
+          ),
+        );
+      });
 
       when(
         () => mockCallState.localParticipant,
@@ -97,19 +110,17 @@ class _CallContentExample extends StatelessWidget {
       child: TestWrapper(
         child: StreamCallContent(
           call: mockCall,
-          callState: mockCallState,
           extendBody: extendBody,
-          callAppBarBuilder: (context, call, callState) => AppBar(
+          callAppBarWidgetBuilder: (context, call) => AppBar(
             title: const Text('Custom App Bar'),
           ),
-          callParticipantsBuilder: (context, call, callState) =>
-              const ColoredBox(
-                color: Colors.green,
-                child: Center(
-                  child: Text('Custom Participants View'),
-                ),
-              ),
-          callControlsBuilder: (context, call, callState) => Row(
+          callParticipantsWidgetBuilder: (context, call) => const ColoredBox(
+            color: Colors.green,
+            child: Center(
+              child: Text('Custom Participants View'),
+            ),
+          ),
+          callControlsWidgetBuilder: (context, call) => Row(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               IconButton(